server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adk-sim-server"
version = "0.1.0"
description = "ADK Agent Simulator gRPC Server"
requires-python = ">=3.14"
dependencies = [
  "adk-sim-protos",
  "grpclib>=0.4.7",
  "sqlalchemy>=2.0.0",
  "databases[aiosqlite]>=0.9.0",
  "python-dotenv>=1.0.1",
  "structlog>=25.5.0",
  "tenacity>=9.1.2",
  "broadcaster>=0.3.1",
]

[dependency-groups]
dev = [
  "adk-sim-testing",
  "pytest>=9.0.1",
  "pytest-asyncio>=1.3.0",
  "pytest-docker>=3.0.0",
  "pytest-timeout>=2.4.0",
  "pyhamcrest>=2.1.0",
  "pyright>=1.1.407",
]

[tool.uv.sources]
adk-sim-protos = { workspace = true }
adk-sim-testing = { workspace = true }

[project.scripts]
adk-sim-server = "adk_sim_server.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_sim_server"]
